Move window class to gui.h.
